agent-assembler/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.0.0
google-cloud-speech>=2.23.0